+++ business/src/main/java/in/koreatech/business/BusinessMainActivity.kt
@@ -11,7 +11,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import dagger.hilt.android.AndroidEntryPoint
-import `in`.koreatech.business.feature.store.MyStoreDetailScreen
+import `in`.koreatech.business.feature.store.storedetail.MyStoreDetailScreen
 import `in`.koreatech.business.ui.theme.KOIN_ANDROIDTheme
 
 @AndroidEntryPoint
