+++ business/src/main/java/in/koreatech/business/feature/store/storedetail/MyStoreDetailScreen.kt
@@ -1,4 +1,4 @@
-package `in`.koreatech.business.feature.store
+package `in`.koreatech.business.feature.store.storedetail
 
 import android.annotation.SuppressLint
 import androidx.compose.foundation.ExperimentalFoundationApi
@@ -47,6 +47,8 @@ import androidx.compose.ui.unit.sp
 import androidx.compose.ui.zIndex
 import androidx.hilt.navigation.compose.hiltViewModel
 import `in`.koreatech.business.R
+import `in`.koreatech.business.feature.store.storedetail.event.EventScreen
+import `in`.koreatech.business.feature.store.storedetail.menu.MenuScreen
 import `in`.koreatech.business.ui.theme.Blue2
 import `in`.koreatech.business.ui.theme.ColorPrimary
 import `in`.koreatech.business.ui.theme.ColorTextField
@@ -237,6 +239,7 @@ fun MyStoreScrollScreen(
                                 isCollapsed,
                                 pagerState.currentPage,
                                 viewModel,
+                                state,
                             ) {
                                 onDeleteEvent()
                             }
