+++ business/src/main/java/in/koreatech/business/feature/store/storedetail/event/EventItem.kt
@@ -1,4 +1,4 @@
-package `in`.koreatech.business.feature.store
+package `in`.koreatech.business.feature.store.storedetail.event
 
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.Image
@@ -9,7 +9,6 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
@@ -20,24 +19,16 @@ import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.pager.HorizontalPager
 import androidx.compose.foundation.pager.PagerState
 import androidx.compose.foundation.pager.rememberPagerState
-import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.Button
-import androidx.compose.material.ButtonDefaults
 import androidx.compose.material.Divider
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.derivedStateOf
-import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
-import androidx.compose.ui.Alignment.Companion.TopStart
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.draw.rotate
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
@@ -45,58 +36,17 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.hilt.navigation.compose.hiltViewModel
 import coil.compose.rememberAsyncImagePainter
 import `in`.koreatech.business.R
+import `in`.koreatech.business.feature.store.storedetail.MyStoreDetailState
+import `in`.koreatech.business.feature.store.storedetail.MyStoreDetailViewModel
 import `in`.koreatech.business.ui.theme.ColorTextField
 import `in`.koreatech.business.ui.theme.Gray1
 import `in`.koreatech.business.ui.theme.Gray2
-import `in`.koreatech.business.ui.theme.Gray4
 import `in`.koreatech.business.ui.theme.Gray6
 import `in`.koreatech.koin.domain.model.store.ShopEvent
-import `in`.koreatech.koin.domain.util.StoreUtil.generateOpenCloseTimeString
-import org.orbitmvi.orbit.compose.collectAsState
+import `in`.koreatech.koin.domain.util.StoreUtil
 
-@Composable
-fun EventScreen(
-    verticalOffset: Boolean,
-    currentPage: Int,
-    viewModel: MyStoreDetailViewModel,
-    onDeleteEvent: () -> Unit
-) {
-    val state = viewModel.collectAsState().value
-    val scrollState = rememberScrollState()
-    val enabledScroll by remember(
-        verticalOffset, scrollState.value
-    ) { derivedStateOf { verticalOffset || scrollState.value != 0 } }
-
-    LaunchedEffect(scrollState.value) {
-        if (scrollState.value != 0 && currentPage != 1) {
-            scrollState.scrollTo(0)
-        }
-    }
-
-    LaunchedEffect(state.storeEvent) {
-        state.isEventExpanded.forEachIndexed { index, item ->
-            viewModel.initEventItem()
-        }
-    }
-    if (state.isEditMode) {
-        EventEditToolBar()
-    } else {
-        EventToolBar()
-    }
-    EventItem(enabledScroll, scrollState, viewModel)
-    OwnerStoreDialog(
-        onDismissRequest = { viewModel.changeDialogVisibility() },
-        onConfirmation = { onDeleteEvent() },
-        dialogTitle = stringResource(R.string.event_delete_title),
-        dialogText = stringResource(R.string.event_delete_text),
-        positiveButtonText = stringResource(id = R.string.delete),
-        visibility = state.dialogVisibility
-    )
-
-}
 
 @OptIn(ExperimentalFoundationApi::class)
 @Composable
@@ -104,8 +54,8 @@ fun EventItem(
     enabledScroll: Boolean,
     scrollState: ScrollState,
     viewModel: MyStoreDetailViewModel,
+    state: MyStoreDetailState,
 ) {
-    val state = viewModel.collectAsState().value
     Column(
         modifier = Modifier
             .fillMaxSize()
@@ -114,25 +64,26 @@ fun EventItem(
     ) {
         state.storeEvent?.forEachIndexed { index, item ->
             val eventOpenCloseTime = remember(item.startDate, item.endDate) {
-                generateOpenCloseTimeString(item.startDate, item.endDate)
+                StoreUtil.generateOpenCloseTimeString(item.startDate, item.endDate)
             }
             val pagerState =
                 rememberPagerState { state.storeEvent[index].thumbnailImages?.size ?: 1 }
             if (state.isEventExpanded[index]) {
                 EventExpandedItem(state.storeEvent[index],
-                    eventOpenCloseTime,
-                    pagerState,
+                    eventOpenCloseTime = eventOpenCloseTime,
+                    pagerState = pagerState,
                     onCollapse = { viewModel.toggleEventItem(index) })
             } else {
                 EventFoldedItem(
-                    state.storeEvent[index],
-                    eventOpenCloseTime,
+                    item = state.storeEvent[index],
+                    eventOpenCloseTime = eventOpenCloseTime,
                     onClicked = {
                         if (!state.isEditMode) viewModel.toggleEventItem(index) else viewModel.onChangeEventSelected(
                             item.eventId
                         )
                     },
                     viewModel = viewModel,
+                    state = state
                 )
             }
             Divider(
@@ -147,141 +98,13 @@ fun EventItem(
 }
 
 @Composable
-fun EventEditToolBar() {
-    val viewModel: MyStoreDetailViewModel = hiltViewModel()
-    val state: MyStoreDetailState = viewModel.collectAsState().value
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(52.dp)
-            .background(Gray4),
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Column(
-            modifier = Modifier
-                .padding(8.dp),
-            verticalArrangement = Arrangement.spacedBy((-6).dp, Alignment.CenterVertically),
-            horizontalAlignment = Alignment.CenterHorizontally,
-        ) {
-            Image(painter = if (state.isAllEventSelected) painterResource(id = R.drawable.ic_check_selected) else painterResource(
-                id = R.drawable.ic_check
-            ),
-                contentDescription = stringResource(R.string.check),
-                modifier = Modifier.clickable { viewModel.onChangeAllEventSelected() })
-            Text(text = stringResource(R.string.all), color = Gray6, fontSize = 12.sp)
-        }
-        Row(
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.Center,
-        ) {
-            Button(
-                onClick = { if (state.isSelectedEvent.size > 1) viewModel.modifyEventError() else viewModel.navigateToModifyScreen() },
-                modifier = Modifier
-                    .width(100.dp)
-                    .padding(8.dp),
-                colors = ButtonDefaults.buttonColors(
-                    backgroundColor = ColorTextField, contentColor = Gray6
-                )
-            ) {
-                Image(
-                    painter = painterResource(id = R.drawable.ic_edit),
-                    contentDescription = stringResource(R.string.modify)
-                )
-                Spacer(modifier = Modifier.width(8.dp))
-                Text(text = stringResource(R.string.modify))
-            }
-            Button(
-                onClick = { viewModel.changeDialogVisibility() },
-                modifier = Modifier
-                    .width(100.dp)
-                    .padding(8.dp),
-                colors = ButtonDefaults.buttonColors(
-                    backgroundColor = ColorTextField, contentColor = Gray6
-                )
-            ) {
-                Image(
-                    painter = painterResource(id = R.drawable.ic_delete),
-                    contentDescription = stringResource(R.string.delete)
-                )
-                Spacer(modifier = Modifier.width(8.dp))
-                Text(text = stringResource(R.string.delete))
-            }
-            Button(
-                onClick = { viewModel.onChangeEditMode() },
-                modifier = Modifier
-                    .width(100.dp)
-                    .padding(8.dp),
-                colors = ButtonDefaults.buttonColors(
-                    backgroundColor = ColorTextField, contentColor = Gray6
-                )
-            ) {
-                Image(
-                    painter = painterResource(id = R.drawable.ic_complete),
-                    contentDescription = stringResource(R.string.complete)
-                )
-                Spacer(modifier = Modifier.width(8.dp))
-                Text(text = stringResource(R.string.complete))
-            }
-        }
-    }
-    Divider(
-        color = Gray4, modifier = Modifier.height(1.dp)
-    )
-}
-
-@Composable
-fun EventToolBar() {
-    val viewModel: MyStoreDetailViewModel = hiltViewModel()
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(52.dp)
-    ) {
-        Button(
-            onClick = { viewModel.onChangeEditMode() },
-            modifier = Modifier
-                .weight(1f)
-                .padding(8.dp),
-            colors = ButtonDefaults.buttonColors(
-                backgroundColor = ColorTextField, contentColor = Color.Black
-            )
-        ) {
-            Image(
-                painter = painterResource(id = R.drawable.ic_edit),
-                contentDescription = stringResource(R.string.edit)
-            )
-            Spacer(modifier = Modifier.width(8.dp))
-            Text(text = stringResource(R.string.edit))
-        }
-        Button(
-            onClick = { /*TODO*/ },
-            modifier = Modifier
-                .weight(1f)
-                .padding(8.dp),
-            colors = ButtonDefaults.buttonColors(
-                backgroundColor = ColorTextField, contentColor = Color.Black
-            )
-        ) {
-            Image(
-                painter = painterResource(id = R.drawable.ic_add_box),
-                contentDescription = stringResource(R.string.add)
-            )
-            Spacer(modifier = Modifier.width(8.dp))
-            Text(text = stringResource(R.string.add))
-        }
-    }
-}
-
-
-@Composable
 fun EventFoldedItem(
     item: ShopEvent,
     eventOpenCloseTime: String,
     onClicked: () -> Unit = {},
-    viewModel: MyStoreDetailViewModel
+    viewModel: MyStoreDetailViewModel,
+    state: MyStoreDetailState,
 ) {
-    val state = viewModel.collectAsState().value
     Row(
         modifier = Modifier
             .clickable(onClick = { onClicked() })
@@ -305,7 +128,7 @@ fun EventFoldedItem(
                         .height(80.dp)
                         .clip(RoundedCornerShape(5.dp)),
                     contentScale = ContentScale.FillBounds,
-                    painter = if (item.thumbnailImages?.size != 0) painterResource(id = R.drawable.no_image) else rememberAsyncImagePainter(
+                    painter = if (item.thumbnailImages?.size == 0) painterResource(id = R.drawable.no_image) else rememberAsyncImagePainter(
                         model = item.thumbnailImages?.getOrNull(0)
                     ),
                     contentDescription = stringResource(R.string.event_default_image),
@@ -314,7 +137,7 @@ fun EventFoldedItem(
                     viewModel.initEventItem()
                     Image(
                         modifier = Modifier
-                            .align(TopStart)
+                            .align(Alignment.TopStart)
                             .height(24.dp)
                             .width(24.dp)
                             .clickable {
