+++ business/src/main/java/in/koreatech/business/feature/store/storedetail/menu/MenuScreen.kt
@@ -1,22 +1,16 @@
-package `in`.koreatech.business.feature.store
+package `in`.koreatech.business.feature.store.storedetail.menu
 
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.Divider
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
@@ -25,20 +19,15 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.hilt.navigation.compose.hiltViewModel
 import `in`.koreatech.business.R
-import `in`.koreatech.business.ui.theme.ColorCategory
-import `in`.koreatech.business.ui.theme.ColorPrimary
-import `in`.koreatech.business.ui.theme.ColorTextField
+import `in`.koreatech.business.feature.store.storedetail.MyStoreDetailState
 import `in`.koreatech.business.ui.theme.Gray3
 import `in`.koreatech.business.ui.theme.Gray6
-import org.orbitmvi.orbit.compose.collectAsState
 
 
 @Composable
@@ -85,7 +74,6 @@ fun MenuScreen(verticalOffset: Boolean, currentPage: Int, state: MyStoreDetailSt
             }
         }
     }
-
     Column(
         modifier = Modifier
             .fillMaxSize()
@@ -93,59 +81,8 @@ fun MenuScreen(verticalOffset: Boolean, currentPage: Int, state: MyStoreDetailSt
     ) {
 
         state.storeMenu?.forEach {
-            Row(
-                modifier = Modifier
-                    .padding(horizontal = 25.dp)
-                    .fillMaxWidth(),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Image(
-                    modifier = Modifier.size(20.dp), painter = painterResource(
-                        id = when (it.id) {
-                            1 -> R.drawable.ic_recommend
-                            2 -> R.drawable.ic_main
-                            3 -> R.drawable.ic_set
-                            4 -> R.drawable.ic_side
-                            else -> R.drawable.ic_recommend
-                        }
-                    ), contentDescription = stringResource(R.string.category)
-                )
-                Text(
-                    text = it.name ?: "",
-                    modifier = Modifier.padding(10.dp),
-                    color = ColorCategory,
-                    fontWeight = FontWeight(500),
-                    fontSize = 18.sp
-                )
-            }
-            it.menus?.forEach { item ->
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(horizontal = 25.dp, vertical = 10.dp),
-                    horizontalArrangement = Arrangement.SpaceBetween,
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Column {
-                        Text(text = item.name, fontWeight = FontWeight(500))
-                        Text(text = item.singlePrice.toString() + "원", color = ColorPrimary)
-                    }
-                    Image(
-                        modifier = Modifier
-                            .width(68.dp)
-                            .height(68.dp),
-                        painter = painterResource(id = R.drawable.ic_koin_logo),
-                        contentDescription = stringResource(R.string.menu_default_image),
-                    )
-                }
-                Divider(
-                    color = ColorTextField,
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(horizontal = 25.dp)
-                        .height(1.dp)
-                )
-            }
+            MenuCategories(it)
+            MenuItem(it)
         }
     }
 }
